fix: オンボーディング機能ボタンの配置変更
